# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface Event {
  id: number;
  name: string;
  dateTime: string;
  location: string;
  createdAt: string;
}

export interface CreateEventRequest {
  name: string;
  dateTime: string;
  location: string;
}

export interface Staff {
  id: number;
  name: string;
  barcodeId: string | null;
  department: string | null;
}

export interface ScanRequest {
  barcodeId: string;
  eventId: number;
}

export type ScanResultStatus =
  (typeof ScanResultStatus)[keyof typeof ScanResultStatus];

export const ScanResultStatus = {
  found: "found",
  not_found: "not_found",
  already_checked_in: "already_checked_in",
  recorded: "recorded",
} as const;

export interface AttendanceRecord {
  id: number;
  staffId: number;
  eventId: number;
  staffName: string;
  checkedInAt: string;
}

export interface ScanResult {
  status: ScanResultStatus;
  staff?: Staff | null;
  attendance?: AttendanceRecord | null;
  message: string;
}

export interface AssignBarcodeRequest {
  barcodeId: string;
}

export interface RecordAttendanceRequest {
  staffId: number;
  eventId: number;
}