# lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceRecord {
  id: number;
  staffId: number;
  eventId: number;
  staffName: string;
  checkedInAt: string;
}