# lib/api-zod/src/generated/types/createEventRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventRequest {
  name: string;
  dateTime: string;
  location: string;
}