# lib/api-zod/src/generated/types/recordAttendanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordAttendanceRequest {
  staffId: number;
  eventId: number;
}