# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecord } from "./attendanceRecord";
import type { ScanResultStatus } from "./scanResultStatus";
import type { Staff } from "./staff";

export interface ScanResult {
  status: ScanResultStatus;
  staff?: Staff | null;
  attendance?: AttendanceRecord | null;
  message: string;
}