# lib/api-zod/src/generated/types/staff.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Staff {
  id: number;
  name: string;
  barcodeId: string | null;
  department: string | null;
}